+++ desktop/packages/mullvad-vpn/src/main/index.ts
@@ -1,5 +1,5 @@
 import { exec, execFile } from 'child_process';
-import { app, nativeTheme, powerMonitor, session, shell, systemPreferences } from 'electron';
+import { app, powerMonitor, session, shell, systemPreferences } from 'electron';
 import fs from 'fs';
 import * as path from 'path';
 import util from 'util';
@@ -468,14 +468,6 @@ class ApplicationMain
       await this.userInterface?.updateTrayTheme();
 
       this.userInterface?.updateTray(this.account.isLoggedIn(), this.tunnelState.tunnelState);
-
-      if (process.platform === 'win32') {
-        nativeTheme.on('updated', async () => {
-          if (this.settings.gui.monochromaticIcon) {
-            await this.userInterface?.updateTrayTheme();
-          }
-        });
-      }
     });
 
     this.registerIpcListeners();